Make GroundedFunctionAtom tests integration tests
